+++ NEWS.md
@@ -8,6 +8,7 @@
 
 ## Minor changes
 
+- New argument `scales` added in the function `get_y_position()`. If the specified value is "free" or "free_y", then the step increase of y positions will be calculated by plot panels. Note that, using "free" or "free_y" gives the same result. A global step increase is computed when scales = "fixed" (#56).
 
 ## Bug fixes